feature(new post): implement redux for new post
- add new post action creator
- update reducer
- add connect
- update Post Form
- update readme

[Finishes #015]
